Replace the immature wildcard matching algorithem with a great one
It's been said the KMP-DFA algorithem is very good at substring
searching yet not support wildcards. So here I tried studying the
underlying technique of the algorithem and successfully managed to
advance it to a new level on which the wildcards, greedy searching,
escape character are fully supported and the compiling and searching
time complexity is considered to remain at O(M+N).